v1: `hamilton`: hold the STAR error definitions once, in `star/driver/errors`
The exception classes, the error-code table and the message tables lived inside
the legacy backend. Moved to pylabrobot/hamilton/star/driver/errors.py, which
the legacy backend now imports from and re-exports, so the v1 driver does not
have to reach into pylabrobot.legacy or carry a second copy.
